change styles on header, page creation and bugfix pagecreation issue
