+++ src/styles/components-style/_formNote.scss
@@ -3,6 +3,7 @@
 @import "../placeholders";
 
 .form-bloc {
+  margin-top: 20px;
   width: 100%;
   max-width: 582px;
   max-height: 242px;
